+++ common/src/main/java/com/coolerpromc/restrictedinventory/network/ServerBoundClientRestrictedSlotsPacket.java
@@ -2,6 +2,7 @@ package com.coolerpromc.restrictedinventory.network;
 
 import com.coolerpromc.restrictedinventory.Constants;
 import com.coolerpromc.restrictedinventory.config.CommonConfig;
+import com.coolerpromc.restrictedinventory.config.util.ItemEntry;
 import com.coolerpromc.restrictedinventory.platform.Services;
 import com.coolerpromc.restrictedinventory.platform.util.PayloadContext;
 import net.minecraft.network.RegistryFriendlyByteBuf;
@@ -12,10 +13,10 @@ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import java.util.HashMap;
 import java.util.Map;
 
-public record ServerBoundClientRestrictedSlotsPacket(Map<Integer, String> restrictedSlots) implements CustomPacketPayload {
+public record ServerBoundClientRestrictedSlotsPacket(Map<Integer, ItemEntry> restrictedSlots) implements CustomPacketPayload {
     public static final Type<ServerBoundClientRestrictedSlotsPacket> TYPE = new Type<>(Constants.id("restricted_slots"));
     public static final StreamCodec<RegistryFriendlyByteBuf, ServerBoundClientRestrictedSlotsPacket> STREAM_CODEC = StreamCodec.composite(
-            ByteBufCodecs.map(HashMap::new, ByteBufCodecs.INT, ByteBufCodecs.STRING_UTF8),
+            ByteBufCodecs.map(HashMap::new, ByteBufCodecs.INT, ItemEntry.STREAM_CODEC),
             ServerBoundClientRestrictedSlotsPacket::restrictedSlots,
             ServerBoundClientRestrictedSlotsPacket::new
     );