+++ common/src/main/java/com/coolerpromc/restrictedinventory/platform/services/IPlatformHelper.java
@@ -1,5 +1,6 @@
 package com.coolerpromc.restrictedinventory.platform.services;
 
+import com.coolerpromc.restrictedinventory.config.util.ItemEntry;
 import net.minecraft.world.entity.player.Player;
 
 import java.util.Map;
@@ -12,8 +13,8 @@ public interface IPlatformHelper {
         return isDevelopmentEnvironment() ? "development" : "production";
     }
     boolean isClient();
-    Map<Integer, String> getRestrictedSlots(Player player);
-    void setRestrictedSlots(Player player, Map<Integer, String> restrictedSlots);
-    void syncRestrictedSlots(Map<Integer, String> restrictedSlots);
+    Map<Integer, ItemEntry> getRestrictedSlots(Player player);
+    void setRestrictedSlots(Player player, Map<Integer, ItemEntry> restrictedSlots);
+    void syncRestrictedSlots(Map<Integer, ItemEntry> restrictedSlots);
     void updatePlayersPermission();
 }