+++ neoforge/src/main/java/com/coolerpromc/restrictedinventory/network/ClientBoundAttachmentSyncPacket.java
@@ -2,6 +2,7 @@ package com.coolerpromc.restrictedinventory.network;
 
 import com.coolerpromc.restrictedinventory.Constants;
 import com.coolerpromc.restrictedinventory.NeoForgeRestrictedInventory;
+import com.coolerpromc.restrictedinventory.config.util.ItemEntry;
 import com.coolerpromc.restrictedinventory.platform.util.PayloadContext;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
@@ -11,10 +12,10 @@ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import java.util.HashMap;
 import java.util.Map;
 
-public record ClientBoundAttachmentSyncPacket(Map<Integer, String> restrictedSlots) implements CustomPacketPayload {
+public record ClientBoundAttachmentSyncPacket(Map<Integer, ItemEntry> restrictedSlots) implements CustomPacketPayload {
     public static final Type<ClientBoundAttachmentSyncPacket> TYPE = new Type<>(Constants.id("attachment_sync"));
     public static final StreamCodec<RegistryFriendlyByteBuf, ClientBoundAttachmentSyncPacket> STREAM_CODEC = StreamCodec.composite(
-            ByteBufCodecs.map(HashMap::new, ByteBufCodecs.INT, ByteBufCodecs.STRING_UTF8),
+            ByteBufCodecs.map(HashMap::new, ByteBufCodecs.INT, ItemEntry.STREAM_CODEC),
             ClientBoundAttachmentSyncPacket::restrictedSlots,
             ClientBoundAttachmentSyncPacket::new
     );