+++ neoforge/src/main/java/com/coolerpromc/restrictedinventory/platform/NeoForgePlatformHelper.java
@@ -1,10 +1,11 @@
 package com.coolerpromc.restrictedinventory.platform;
 
 import com.coolerpromc.restrictedinventory.NeoForgeRestrictedInventory;
+import com.coolerpromc.restrictedinventory.config.util.ItemEntry;
 import com.coolerpromc.restrictedinventory.network.ClientBoundAttachmentSyncPacket;
 import com.coolerpromc.restrictedinventory.platform.services.IPlatformHelper;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.loading.FMLEnvironment;
@@ -35,12 +36,12 @@ public class NeoForgePlatformHelper implements IPlatformHelper {
     }
 
     @Override
-    public Map<Integer, String> getRestrictedSlots(Player player) {
+    public Map<Integer, ItemEntry> getRestrictedSlots(Player player) {
         return player.getData(NeoForgeRestrictedInventory.RESTRICTED_SLOTS_ATTACHMENT);
     }
 
     @Override
-    public void setRestrictedSlots(Player player, Map<Integer, String> restrictedSlots) {
+    public void setRestrictedSlots(Player player, Map<Integer, ItemEntry> restrictedSlots) {
         player.setData(NeoForgeRestrictedInventory.RESTRICTED_SLOTS_ATTACHMENT, restrictedSlots);
         if (player instanceof ServerPlayer serverPlayer){
             Services.NETWORK.sendToPlayer(serverPlayer, new ClientBoundAttachmentSyncPacket(restrictedSlots));
@@ -48,7 +49,7 @@ public class NeoForgePlatformHelper implements IPlatformHelper {
     }
 
     @Override
-    public void syncRestrictedSlots(Map<Integer, String> restrictedSlots) {
+    public void syncRestrictedSlots(Map<Integer, ItemEntry> restrictedSlots) {
         if (ServerLifecycleHooks.getCurrentServer() != null){
             ServerLifecycleHooks.getCurrentServer().getPlayerList().getPlayers().forEach(p -> setRestrictedSlots(p, restrictedSlots));
         }